Tidying compile errors
